Move deployment classes to a deployment package
This is the convention and avoids having a split package between the
runtime artifact and the deployment one.
